+++ src/connector.rs
@@ -39,9 +39,8 @@ impl Builder {
     }
 
     /// Build a HTTP connector
-    #[cfg(feature = "tcp")]
-    pub fn build(self) -> hyper::client::connect::HttpConnector {
-        hyper::client::connect::HttpConnector::new()
+    pub fn build(self) -> hyper_util::client::legacy::connect::HttpConnector {
+        hyper_util::client::legacy::connect::HttpConnector::new()
     }
 }
 
@@ -96,7 +95,9 @@ impl HttpsBuilder {
     pub fn build(
         self,
     ) -> Result<
-        hyper_openssl::HttpsConnector<hyper::client::HttpConnector>,
+        hyper_openssl::client::legacy::HttpsConnector<
+            hyper_util::client::legacy::connect::HttpConnector,
+        >,
         openssl::error::ErrorStack,
     > {
         // SSL implementation
@@ -114,20 +115,23 @@ impl HttpsBuilder {
             ssl.check_private_key()?;
         }
 
-        let mut connector = hyper::client::HttpConnector::new();
+        let mut connector = hyper_util::client::legacy::connect::HttpConnector::new();
         connector.enforce_http(false);
-        hyper_openssl::HttpsConnector::<hyper::client::HttpConnector>::with_connector(
-            connector, ssl,
-        )
+        hyper_openssl::client::legacy::HttpsConnector::<
+            hyper_util::client::legacy::connect::HttpConnector,
+        >::with_connector(connector, ssl)
     }
 
     #[cfg(any(target_os = "macos", target_os = "windows", target_os = "ios"))]
     /// Build the HTTPS connector. Will fail if the SSL connector can't be created.
     pub fn build(
         self,
-    ) -> Result<hyper_tls::HttpsConnector<hyper::client::HttpConnector>, native_tls::Error> {
+    ) -> Result<
+        hyper_tls::HttpsConnector<hyper_util::client::legacy::connect::HttpConnector>,
+        native_tls::Error,
+    > {
         let tls = native_tls::TlsConnector::new()?.into();
-        let mut connector = hyper::client::HttpConnector::new();
+        let mut connector = hyper_util::client::legacy::connect::HttpConnector::new();
         connector.enforce_http(false);
         let mut connector = hyper_tls::HttpsConnector::from((connector, tls));
         connector.https_only(true);