+++ src/multipart/related.rs
@@ -1,7 +1,8 @@
 //! Helper functions for multipart/related support
 
-use hyper_0_10::header::{ContentType, Headers};
-use mime_0_2::Mime;
+use hyper::header::{HeaderValue, CONTENT_TYPE};
+use hyper::HeaderMap;
+use mime::Mime;
 
 /// Construct the Body for a multipart/related request. The mime 0.2.6 library
 /// does not parse quoted-string parameters correctly. The boundary doesn't
@@ -20,19 +21,20 @@ pub fn generate_boundary() -> Vec<u8> {
 
 /// Create the multipart headers from a request so that we can parse the
 /// body using `mime_multipart::read_multipart_body`.
-pub fn create_multipart_headers(
-    content_type: Option<&hyper::header::HeaderValue>,
-) -> Result<Headers, String> {
+pub fn create_multipart_headers(content_type: Option<&HeaderValue>) -> Result<HeaderMap, String> {
     let content_type = content_type
         .ok_or_else(|| "Missing Content-Type header".to_string())?
         .to_str()
         .map_err(|e| format!("Couldn't read Content-Type header value: {}", e))?
         .parse::<Mime>()
         .map_err(|_e| "Couldn't parse Content-Type header value".to_string())?;
-
     // Insert top-level content type header into a Headers object.
-    let mut multipart_headers = Headers::new();
-    multipart_headers.set(ContentType(content_type));
+    let mut multipart_headers = HeaderMap::new();
+    multipart_headers.append(
+        CONTENT_TYPE,
+        HeaderValue::from_str(&content_type.to_string())
+            .map_err(|_e| "Couldn't parse Content-Type header value".to_string())?,
+    );
 
     Ok(multipart_headers)
 }