+++ CHANGELOG.md
@@ -5,7 +5,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Changed
 
 - *BREAKING* - Remove dependency and re-export of `hyper-old-types` which is no longer maintained.
@@ -13,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
       - The `hyper_old_types` are no longer re-exported, and the enums just wrap `String`s.
       - `AuthData::bearer()` now returns `Option`, returning `None` if the provided token is not valid base64.
   - The `auth::make_headers` function now returns an `Option<AuthData>`.
+- *BREAKING* - Update to hyper 1.x
+  - The `tcp` feature has been removed, to match the new hyper feature set.
+- *BREAKING* - Remove `SwaggerService` trait due to lack of usage.
 
 ### Added